add reset channel and fix purge

+++ commands/purge.js
@@ -5,6 +5,9 @@ module.exports = {
   execute(message, number) {
 
     let num = parseInt(number);
+
+    if (num > 100) return message.reply("Please enter a number less than or equal to 100");
+    
     message.channel.messages.fetch({ limit: num}).then(messages => {
       message.channel.bulkDelete(messages).then(deleted => {
         message.reply(`${deleted.size} messages have been successfully deleted!`).then(message => message.delete(5000));
